change int to long for accumulated, Normalize, Denormalize, TryReadValue/WriteValue, GetRequiredBufferSize; use ulong for zigzag arithmetic; update block max length to 13
Agent-Logs-Url: https://github.com/petesramek/polyline-algorithm-csharp/sessions/b16b1bac-62fc-44de-9c08-88c566cf476e

+++ src/PolylineAlgorithm/PolylineEncoder.cs
@@ -110,7 +110,7 @@ public class PolylineEncoder<TCoordinate, TPolyline> : IPolylineEncoder<TCoordin
                     long delta = current - previous[j];
                     previous[j] = current;
 
-                    if (!PolylineEncoding.TryWriteValue((int)delta, buffer, ref position)) {
+                    if (!PolylineEncoding.TryWriteValue(delta, buffer, ref position)) {
                         _logger.LogOperationFailedDebug(OperationName);
                         _logger.LogCannotWriteValueToBufferWarning(position, i);
                         ExceptionGuard.ThrowCouldNotWriteEncodedValueToBuffer();
